feat: AudioManager question progress framework
Built a structure to deal with the progression through the piece. Right now it is extremely straight-forward (problematically so), but it's a place to start. Each question triggers a case in a switch statement, changing the audio accordingly. In the future, we will need to deal with:

* the possibility of branching
* the possibility that some branches won't be as long as others
* a more complex "possibility-space" in general

the next update will add a rough sample of what this musical walkthrough might entail (warning: it will be verrrrrry heavy-handed. Subtlety is hard won.