feat: Progression Workshop
Here we have an example (in build form) of the progression. It is not very good and far too heavy-handed (as promised), but also (perhaps not surprisingly) not obvious enough. This is subtle work, y'all!

In this build, note that:
* the first question picks the lead instrument based on your answer (0-3 = piano / 4-6 = vocal / 7-9 = guitar) This is fine, but could be better obviously. But the framework is there
* the remaining questions set an effect level based on whatever number you pick (echo mix, echo amount, distortion level, and pitch)
* everything is really rough and not nuanced, especially the distortion control

While experimenting with the effect levels (in particular the pitch effects), it becomes fun to hear how the pitch and echo's work together to create more complex trails. This could be a fun thing to try, i.e. a T/F question that starts a pitch sweep? Could be applied to other effect sweeps as well (mainly, filter sweeps)