feat: Question Management
Okay. So here's an interesting commmit. Here we have the start of a system for building out a template for questions. These two question types (slider and short answer) builds on a common Question script. This common Question script has places for setting the type of effect to be fired, and some other stuff, which can be added to as the project moves forward.

The questions as we move forward seem to be these:
* What general effects do we want to see happen?
* Do we want these musical effects to be fired right away, or is there some benefit to having them build up and then get released all at once? If this was the case, there could be a place (for example) to note tempo, key signature, instrumentation, of a given element and then begin it when these things are sorted
* What types of questions do we want to ask? Right now we have short answer and slider, but there is definitely space for more varied question (and activity) types

The best course seems to just start making and not worry too much about the framework (as tempting as that is to fret over)