test: Versioning Ableton
Interesting. it does not let us specifically version the .als file OR the .cfg file (at least as far as can be seen with this initial test. Strangely enough, the files have been uploaded to GitHub so you would assume that they could be versioned, but it won't actually pull the .cfg file on a pull request, so... not sure what's going on there)

However, it does let us include these project backup files that recall an earlier state of the project. If we directly name, date, and upload the files in a searchable way, this might be a great way to do this. Test further.