feat: recorded output working
Yes! Need to make sure that this will save the file in the build. Also need to make sure that we don't need specialized naming in case the player plays the game more than once (actually, probably better to be safe rather than sorry and just go ahead and build it that way for the next update...)