process: fixing Recording and Build
Finally got the build working so that the audio is being played from PD and the Recording Output happens. The RecordAudio function also now generates a random number to attach to the file so we don't run up against the file management issues from earlier.

Note: for the PD problem, the issue was that within PD the files were relationally linked, but that relationship was changing once the build happens. For the future, remember that audio files need to be stored in the streaming assets folder (along with the PD patches) so that they can be accessed. bboysil's ConsoleToGUI script from the Unity Forums (https://answers.unity.com/questions/125049/is-there-any-way-to-view-the-console-in-a-build.html) was invaluable to discover this, because I was able to check the "pipe PD print to Console" option and actually see what was hapennign within the patches

Next:
* work on rearranging the questions
* narrative (or informational) texts iin place of the second books question
* create three (at least) different ways the composition could end
* tweak the beginning so there is more variety
* slow down speed of chimers in the intro
* reduce volume of chimers when the keys come in