feat: Creep!
Now the program can get the player's IP address, and from that get the current weather. IN this version the weather is applied to the initial overall pitch. It's mapped to the coldest and warmest places on earth (Oymyakon, Siberia @ 202Kelvin and Death Valley, CA @ 327Kelvin) but that's probably too broad a range given that most players won't be at those extremes.

From this, we also have access to a ton of other info, most easily used perhaps being the city. "I notice you're in Montreal. Is that a nice place to be?" etc.