feat: New UI
spent the last day changing the UI away from the basic Unity slider that was originally implemented. The new interface puts the options down in an array at the bottom of the screen. Briefly thought about making them more "button-like" or stacked in a list (a la Disco Elysium dialogue choices) but I kind of like the way it is now. Will sit with it for a bit.

Thoughts/Next:
* update the banana dialogue so it feels a little more branching. Maybe add more of those branches (but would need to first figure out a framework update to deal with those branches)
* likewise, is there other ways we can lean on openAI here. Specifically, it could be fun if the AI reads you a story after the book question. Maybe based on your current city
* Work with the audio to be more dynamic (most notably change the echoing pings in the beginning as that becomes too constant for all playtests)